@@ -5,6 +5,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 
 namespace Estimate
 {
@@ -34,7 +35,7 @@ namespace Estimate
                     builder.AllowAnyHeader().AllowAnyMethod().AllowAnyOrigin().AllowCredentials();
                 });
             });
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_1);
+            services.AddMvc(options => { options.EnableEndpointRouting = false; }).SetCompatibilityVersion(CompatibilityVersion.Version_3_0);
             services.AddSignalR(config =>
             {
                 config.EnableDetailedErrors = true;
@@ -42,7 +43,7 @@ namespace Estimate
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
@@ -62,8 +63,7 @@ namespace Estimate
             app.UseCors();
             app.UseMvc();
 
-            app.UseSignalR(config =>
-            {
+            app.UseEndpoints(config => {
                 config.MapHub<EstimateHub>("/estimate");
             });
         }
